+++ docs/versionhistory.rst
@@ -58,8 +58,12 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
   ``Semaphore.acquire_nowait()`` methods
 - Added the ``statistics()`` method to ``Event``, ``Lock``, ``Condition``, ``Semaphore``,
   ``CapacityLimiter``, ``MemoryObjectReceiveStream`` and ``MemoryObjectSendStream``
-- Added the ``anyio.lowlevel`` module containing the ``checkpoint()``,
-  ``checkpoint_if_cancelled()`` and ``cancel_shielded_checkpoint()`` functions
+- Added the ``anyio.lowlevel`` module containing:
+
+  * The ``checkpoint()`` function
+  * The ``checkpoint_if_cancelled()`` function
+  * The ``cancel_shielded_checkpoint()`` function
+  * The ``RunVar()`` class
 - Changed ``CancelScope.deadline`` to be writable
 - ``Lock`` and ``Condition`` can now only be released by the task that acquired them. This behavior
   is now consistent on all backends whereas previously only Trio enforced this.